requirements.txt
openvino-dev==2023.0.0
click